requirements.txt
altair==5.5.0
annotated-types==0.7.0
attrs==25.3.0
awscli==1.38.35
beautifulsoup4==4.12.3
blinker==1.9.0
boto3==1.37.11
botocore==1.37.35
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dash==3.0.3
dash-bootstrap-components==2.0.2
docutils==0.19
Flask==3.0.3
fonttools==4.56.0
frozendict==2.4.6
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
gunicorn==23.0.0
html5lib==1.1
httplib2==0.22.0
humanize==4.12.2
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lxml==5.3.0
MarkupSafe==3.0.2
matlib==0.1.2
matplotlib==3.10.1
mplcursors==0.6
multitasking==0.0.11
narwhals==1.35.0
nest-asyncio==1.6.0
numpy==2.2.5
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
peewee==3.17.8
pillow==11.2.1
platformdirs==4.3.6
plotly==6.0.1
pluggy==1.5.0
proto-plus==1.26.1
protobuf==5.29.4
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.5
pydantic_core==2.27.2
pydeck==0.9.1
pyparsing==3.2.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rpds-py==0.24.0
rsa==4.7.2
s3transfer==0.11.4
seaborn==0.13.2
setuptools==79.0.0
six==1.17.0
smmap==5.0.2
soupsieve==2.6
streamlit==1.44.1
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
waitress==3.0.2
watchdog==6.0.0
webencodings==0.5.1
Werkzeug==3.0.6
yfinance==0.2.54
zipp==3.21.0
